Add therubyracer or bundler breaks $LOAD_PATH
Bundler prunes Ruby's $LOAD_PATH down to only gems that are listed in
`Gemfile` and their dependencies. When a package does not explicitly depend on
a particular gem, as *execjs* does because it is able to use any of
several JavaScript engines, it gets removed from the load path. I have
been unable to find a way to configure bundler to not do this. The other
alternative would be to remove the Gemfile altogether (or maybe rename
it to something not default).